cloudeos/module: Add routing_resource_info to router_status
routing_resource_info is a list of all the route table and
route_table_associations resources created in the route table module.
This will add an explicit dependency of these resources on
cloudeos_router_status.

The reason for this dependency is that during a delete, the
route tables and associations are deleted before the router_status
is deleted and that causes the CloudEOS router to be unreachable.
Which further causes tasks to timeout.

Change-Id: Ia84195dfa03517071c9cf56a83f2eab26a09142a