terraform/examples: Refactor hardcoded values
The example topologies have hardcoded values for cidr blocks, subnets
and IPs. Due to this, it is difficult to change them without modifying
the entire terraform file.

This fix replaces the hardcoded values in aws_tworegion_noleaf topology
with variables. This is needed while implementing internal cluster
routing which overrides these values with new ones.

Closes: BUG600385
Change-Id: I237ab23c0db205822e58af5f741c0cc14c18591b