terraform/examples: Modify input variables in aws_tworegion_noleaf
Presently, the variables used in the no leaf topology do not follow
any convention. Thus, if we want to replace them later on during
internal cluster deployment(cicada), it would be error-prone.

This change clubs all the variables in json like hierarchical format
which can be parsed and replaced easily.

Change-Id: Ibd9a94ec42a6d9be4686c89519ae81c09798a8b6
Signed-off-by: manishp <manishp@arista.com>