terraform/examples: Refactor hardcoded values for aws_oneregion_multipleleaf
The example topologies have hardcoded values for cidr blocks, subnets
and IPs. Due to this, it is difficult to change them without modifying
the entire terraform file.

This fix replaces the hardcoded values in aws_oneregion_multipleleaf topology
with variables. This is needed while implementing internal cluster
routing which overrides these values with new ones.

Change-Id: I27ed9b5919e7437e88008c2a43a125e85f3a8c58